+++ src/App/App.css
@@ -1,5 +1,5 @@
 body {
-  background-image: url('inkblots.jpeg');
+  background-image: url('../inkblots.jpeg');
   background-size: cover;
 }
